Revise tests to use committed test file.
